# apps/api/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "RuleMind API",
    "version": "0.1.0"
  },
...
  "paths": {
...
    "/api/v1/approvals/{id}/approve": {
      "post": {
        "parameters": [
          {
            "schema": {
              "type": "string"
            },
            "in": "path",
            "name": "id",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "Default Response"
          }
        }
      }
    },
    "/api/v1/approvals/{id}/reject": {
      "post": {
        "parameters": [
          {
            "schema": {
              "type": "string"
            },
            "in": "path",
            "name": "id",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "Default Response"
          }
        }
      }
    },
...
  }
}